+++ examples/materials.rs
@@ -14,18 +14,18 @@ fn main() {
 
     let geometry = three::Geometry::cylinder(1.0, 2.0, 2.0, 5);
     let mut materials = vec![
-        three::material::MeshBasicBuilder::new().color(0xffffff).build(),
-        three::material::MeshLambertBuilder::new().color(0xffffff).flat(true).build(),
-        three::material::MeshLambertBuilder::new().color(0xffffff).build(),
-        three::material::MeshPhongBuilder::new().color(0xffffff).glossiness(80.0).build(),
-        three::material::MeshPbrBuilder::new()
+        three::Material::mesh_basic().color(0xffffff).build(),
+        three::Material::mesh_lambert().color(0xffffff).flat(true).build(),
+        three::Material::mesh_lambert().color(0xffffff).build(),
+        three::Material::mesh_phong().color(0xffffff).glossiness(80.0).build(),
+        three::Material::mesh_pbr()
         .base_color_factor([0.2,  0.2, 0.2, 1.0])
         .metallic_roughness([0.5, 0.5])
         .occlusion_strength(0.2)
         .normal_scale(1.0)
         .build(),
     ];
-    
+
     let count = materials.len();
 
     let _cubes: Vec<_> = materials.drain(..).enumerate().map(|(i, mat)| {
