+++ src/materials/builder.rs
@@ -1,84 +1,12 @@
-use render::BasicPipelineState;
-use scene::Color;
-use texture::Texture;
-
-/// Material is the enhancement of Texture that is used to setup appearance of [`Mesh`](struct.Mesh.html).
-#[allow(missing_docs)]
-#[derive(Clone, Debug)]
-pub enum Material {
-    /// Basic wireframe with specific `Color`.
-    LineBasic { color: Color },
-    /// Basic material with color, optional `Texture` and optional wireframe mode.
-    MeshBasic {
-        color: Color,
-        map: Option<Texture<[f32; 4]>>,
-        wireframe: bool,
-    },
-    /// Lambertian diffuse reflection. This technique causes all closed polygons
-    /// (such as a triangle within a 3D mesh) to reflect light equally in all
-    /// directions when rendered.
-    MeshLambert { color: Color, flat: bool },
-    /// Material that uses Phong reflection model.
-    MeshPhong { color: Color, glossiness: f32 },
-    /// Physically-based rendering material.
-    MeshPbr {
-        base_color_factor: [f32; 4],
-        metallic_roughness: [f32; 2],
-        occlusion_strength: f32,
-        emissive_factor: [f32; 3],
-        normal_scale: f32,
-        base_color_map: Option<Texture<[f32; 4]>>,
-        normal_map: Option<Texture<[f32; 4]>>,
-        emissive_map: Option<Texture<[f32; 4]>>,
-        metallic_roughness_map: Option<Texture<[f32; 4]>>,
-        occlusion_map: Option<Texture<[f32; 4]>>,
-    },
-    /// 2D Sprite.
-    Sprite { map: Texture<[f32; 4]> },
-    /// Custom material.
-    CustomBasicPipeline {
-        color: Color,
-        map: Option<Texture<[f32; 4]>>,
-        pipeline: BasicPipelineState,
-    },
-}
-
-
-
-
 #[derive(Debug)]
-pub struct LineBasicBuilder {
-    color: u32
-}
-
-impl LineBasicBuilder {
-    pub fn new() -> LineBasicBuilder {
-        LineBasicBuilder{
-            color: 0
-        }
-    }
-    pub fn color(&mut self, color: u32) -> &mut LineBasicBuilder{
-        self.color = color;
-        self
-    }
-
-    pub fn build(&self) -> Material {
-        Material::LineBasic{color: self.color}
-    }
-
-}
-
-#[derive(Debug)]
-
 pub struct MeshBasicBuilder {
     color: u32,
     map: Option<Texture<[f32;4]>>,
     wireframe: bool,
 }
 
-
 impl MeshBasicBuilder {
-    pub fn new() -> MeshBasicBuilder{
+    pub(crate) fn new() -> MeshBasicBuilder{
         MeshBasicBuilder{
             color: 0,
             map: None,
@@ -117,7 +45,7 @@ pub struct MeshLambertBuilder {
 }
 
 impl  MeshLambertBuilder {
-    pub fn new() -> MeshLambertBuilder {
+    pub(crate) fn new() -> MeshLambertBuilder {
         MeshLambertBuilder{
             color: 0,
             flat: false
@@ -150,7 +78,7 @@ pub struct MeshPhongBuilder {
 }
 
 impl MeshPhongBuilder {
-    pub fn new() -> MeshPhongBuilder{
+    pub(crate) fn new() -> MeshPhongBuilder{
         MeshPhongBuilder{
             color:0,
             glossiness: 0.0
@@ -190,7 +118,7 @@ pub struct MeshPbrBuilder {
 }
 
 impl MeshPbrBuilder{
-     pub fn new() -> MeshPbrBuilder {
+     pub(crate) fn new() -> MeshPbrBuilder {
          MeshPbrBuilder{
             base_color_factor: [0.0;4],
             metallic_roughness: [0.0; 2],
@@ -271,4 +199,3 @@ impl MeshPbrBuilder{
      }
      
 }
-
